Add magnetism between the selected piece and the figure to reproduce.

+++ src/Model/Piece.h
@@ -7,8 +7,8 @@
 
 #include <list>
 #include <cstdarg>
-#include <FileUtils.h>
 
+#include "../Controller/FileUtils.h"
 #include "Shape.h"
 #include "Triangle.h"
 
@@ -182,7 +182,7 @@ double Piece<CoordinateType>::distance(Shape<CoordinateType>* shape, std::vector
 template<class CoordinateType>
 Piece<CoordinateType>::Piece(sf::Color color, vector<Point<CoordinateType>> points) {
 	Point<double> offsetPoint(600, 50);
-	for(int i = 0; i < points.size() / 3; i++) {
+	for(int i = 0; i < (int)points.size() / 3; i++) {
 		Point<CoordinateType> p1 = points[i * 3 ] + offsetPoint;
 		Point<CoordinateType> p2 = points[i * 3 + 1] + offsetPoint;
 		Point<CoordinateType> p3 = points[i * 3 + 2] + offsetPoint;
